## What is the purpose of the change
Fix the performance regression described in FLINK-30624.
```
before this fix:
Benchmark                                  (stateBackend)   Mode  Cnt     Score    Error   Units
MemoryStateBackendBenchmark.stateBackends          MEMORY  thrpt   30  3936.158 ± 65.696  ops/ms
MemoryStateBackendBenchmark.stateBackends              FS  thrpt   30  3963.934 ± 64.942  ops/ms
MemoryStateBackendBenchmark.stateBackends        FS_ASYNC  thrpt   30  3930.955 ± 68.809  ops/ms`

after this fix:
Benchmark                                  (stateBackend)   Mode  Cnt     Score    Error   Units
MemoryStateBackendBenchmark.stateBackends          MEMORY  thrpt   30  4151.066 ± 80.526  ops/ms
MemoryStateBackendBenchmark.stateBackends              FS  thrpt   30  4143.619 ± 65.212  ops/ms
MemoryStateBackendBenchmark.stateBackends        FS_ASYNC  thrpt   30  4161.912 ± 44.701  ops/ms
```

## Verifying this change
This change is already covered by existing tests

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): (**no**)
  - The public API, i.e., is any changed class annotated with `@Public(Evolving)`: (**no**)
  - The serializers: (**no**)
  - The runtime per-record code paths (performance sensitive): (**no**)
  - Anything that affects deployment or recovery: JobManager (and its components), Checkpointing, Kubernetes/Yarn, ZooKeeper: (**no**)
  - The S3 file system connector: (**no**)

## Documentation

  - Does this pull request introduce a new feature? (**no**)
  - If yes, how is the feature documented? (**not applicable**)