@wanglijie95 2023-01-30
@flinkbot run azure